backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==11.0.0
python-dotenv==1.0.1
requests==2.31.0
werkzeug==3.0.1
google-generativeai==0.8.3
google-genai==0.3.0